AI/requirements.txt
llama-cpp-python==0.2.11
numpy>=1.21.0